+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - fix workflow
 - add CodeClimate.com, CodeFactor.io, snyk.io
 - refactor convert_1_0000_0000_to_9999_9999_9999_9999
-- refactor CharacterSet
+- refactor CharacterSet, fillZeroIfTrue
 - refactor test into Simplified_Normal, Traditional_Normal, Simplified_Capital
 
 ## `1.6.1` - 18 June 2022